+++ index.js
@@ -73,6 +73,10 @@ function createError () {
     }
   }
 
+  if (typeof status === 'number' && (status < 400 || status >= 600)) {
+    deprecate('non-error status code; use only 4xx or 5xx status codes')
+  }
+
   if (typeof status !== 'number' ||
     (!statuses[status] && (status < 400 || status >= 600))) {
     status = 500